move ast
