+++ pyproject.toml
@@ -242,7 +242,7 @@ convention = "numpy"
 "xonsh/xoreutils/*.py" = ["E722"]
 "xonsh/completers/python.py" = ["E722"]
 "xonsh/ptk/shell.py" = ["E731"]
-"xonsh/ast.py" = ["F401"]
+"xonsh/parser/ast.py" = ["F401"]
 "xonsh/readline_shell.py" = ["F401"]
 "xonsh/commands_cache.py" = ["F841"]
 "xonsh/ptk/key_bindings.py" = ["F841"]
