+++ xonsh/parser/context_check.py
@@ -2,7 +2,7 @@ import ast
 import collections
 import keyword
 
-from xonsh import ast as xast
+from xonsh.parser import ast as xast
 
 _all_keywords = frozenset(keyword.kwlist)
