+++ xonsh/parser/fstring_adaptor.py
@@ -3,7 +3,7 @@
 import re
 from ast import parse as pyparse
 
-from xonsh import ast
+from xonsh.parser import ast
 from xonsh.lazyasd import lazyobject
 from xonsh.platform import PYTHON_VERSION_INFO
